Time: 63 ms (93.4%), Space: 155.6 MB (75.58%) - LeetHub

+++ 3867-sum-of-gcd-of-formed-pairs/3867-sum-of-gcd-of-formed-pairs.cpp
@@ -11,7 +11,7 @@ public:
         int i = 0;
         int j = prefix.size()-1;
         long long sum = 0;
-        int long ans;
+        long long ans;
         while(i<j){
             ans = gcd(prefix[i],prefix[j]);
             sum+=ans;
